+++ tests/testthat/test-buildAchilles.R
@@ -1,9 +1,9 @@
-test_that("test buildAchilles", {
+test_that("test buildAchillesTables", {
 
   cdm <- omock::mockCdmFromDataset() |>
     copyCdm()
 
-  expect_no_error(cdm <- buildAchilles(cdm = cdm, achillesId = "all"))
+  expect_no_error(cdm <- buildAchillesTables(cdm = cdm, achillesId = "all"))
 
   dropCreatedTables(cdm = cdm)
 })