+++ NEWS.md
@@ -2,7 +2,7 @@
 # OmopConstructor 0.3.0
 
 * Added working support for snowflake drivers. by @james-cockayne in #64
-* Create buildAchilles function by @catalamarti in #59
+* Create buildAchillesTables function by @catalamarti in #59
 
 # OmopConstructor 0.2.0